Some page modify

+++ src/Pages/MyCard/MyCard.jsx
@@ -1,22 +1,46 @@
+import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
+import Swal from 'sweetalert2'
 
 const MyCard = () => {
-  const cardWatch = useLoaderData();
+  const loadedCardWatch = useLoaderData();
+  const [cardWatch, setCardWatch] = useState(loadedCardWatch);
 
   const handleDelete = (_id) => {
-    console.log(_id);
-
-    fetch(`http://localhost:3001/card/${_id}`,{
-        method: 'DELETE',
-    })
-    .then(res => res.json())
-    .then(data => console.log(data))
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:3001/card/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount > 0){
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+              const remaining = loadedCardWatch.filter(
+                (watch) => watch._id !== _id
+              );
+              setCardWatch(remaining);
+            }
+          });
+      }
+    });
   };
 
   return (
-    <div>
+    <div className="min-h-screen">
       <div className="container mx-auto px-5 md:px-10 my-20">
-        <h1>My card page</h1>
+        <h1 className="uppercase text-center text-3xl font-bold my-10">
+          My Card Products
+        </h1>
         <div className="overflow-x-auto">
           <table className="table">
             {/* head */}
@@ -37,7 +61,14 @@ const MyCard = () => {
                   <td>{product.name}</td>
                   <td>{product.brand}</td>
                   <td>${product.price}</td>
-                  <td><button onClick={()=>handleDelete(product._id)} className="primary-btn text-white w-32">Delete</button></td>
+                  <td>
+                    <button
+                      onClick={() => handleDelete(product._id)}
+                      className="primary-btn text-white w-32"
+                    >
+                      Delete
+                    </button>
+                  </td>
                 </tr>
               </tbody>
             ))}
